move env var parsing logic to server
